backend/requirements.txt
# ===========================================
# AutoIMS Backend - Python Dependencies
# ===========================================
# Install with: pip install -r requirements.txt
# ===========================================

# --- Core Framework ---
Flask==3.0.0                    # Web framework
Werkzeug==3.0.1                 # WSGI utility library (Flask dependency)

# --- Database ---
Flask-SQLAlchemy==3.1.1         # SQLAlchemy integration for Flask
SQLAlchemy>=2.0.0               # Python SQL toolkit and ORM
psycopg[binary]>=3.1.0          # PostgreSQL adapter (psycopg3)

# --- Authentication & Security ---
PyJWT==2.8.0                    # JSON Web Token implementation

# --- CORS Support ---
Flask-CORS==4.0.0               # Cross-Origin Resource Sharing support

# --- Utilities ---
python-dotenv==1.0.0            # Load environment variables from .env file
